+++ src/main/java/com/alkemy/ong/data/repositories/OrganizationRepository.java
@@ -1,6 +1,6 @@
-package com.alkemy.ong.data.repository;
+package com.alkemy.ong.data.repositories;
 
-import com.alkemy.ong.data.entity.OrganizationEntity;
+import com.alkemy.ong.data.entities.OrganizationEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
